requirements.txt
colorama
tqdm
IPython
matplotlib
joblib
click
numpy
Pillow
natsort
ffmpeg
librosa
